requirements.txt
anthropic
openai
google-generativeai
gigachat
requests
pyyaml
